Refactor Loot Modifiers in their own package
