+++ src/main/java/com/blamejared/crafttweaker/impl/loot/CTLootManager.java
@@ -1,6 +1,7 @@
 package com.blamejared.crafttweaker.impl.loot;
 
 import com.blamejared.crafttweaker.api.annotations.ZenRegister;
+import com.blamejared.crafttweaker.impl.loot.modifier.CTLootModifierManager;
 import com.blamejared.crafttweaker_annotations.annotations.Document;
 import org.openzen.zencode.java.ZenCodeGlobals;
 import org.openzen.zencode.java.ZenCodeType;
